pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Bot that answers form questions using PDF OCR and an LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "seleniumbase",
    "pdf2image",
    "pytesseract",
    "openai",
    "python-dotenv",
    "opencv-python-headless",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "mypy",
]

[project.scripts]
bot = "bot.main:main"

[tool.pytest.ini_options]
markers = ["cloudflare: requires a live browser and network (deselect with '-m not cloudflare')"]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
